+++ net-calculator-tester/Calculator.cs
@@ -13,13 +13,13 @@ namespace net_calculator_tester
         public static float Subtract(float a, float b) { return a - b; }
 
         
-        public static double Divide(double a, double b)
+        public static float Divide(double a, double b)
         {
             if (b == 0)
             {
                 throw new System.ArgumentException("Impossibile dividere per zero");
             }
-            return a / b;
+            return (float)(a / b);
         }
 
         public static float Multiply(float a, float b) { return a*b; }
